+++ addons/easing_curve/scripts/editor/backend/native_curve_editor_backend.gd
@@ -9,7 +9,7 @@ var _control_drag_scale := Vector2.ONE
 
 const IMPLEMENTED_TRANSITION_IDS := [
 	0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
-	100, 101, 102, 103, 104, 105, 106, 107, 108,
+	100, 101, 102, 103, 104, 105, 106, 107, 108, 109,
 ]